style:lint not suggest use ALL_CAPS with const name
